editoast: create object_ref.rs in primitives folder
